lesson4/requirements.txt
web3==5.23.1
py-solc-x
python-dotenv
